Removed unused functions, rename files, and add tests
